# fix: replace godirwalk with filepath.WalkDir